requirements.txt
pypdf2
pdf2image
pillow
google-generativeai
python-docx
flask
werkzeug